# openapi.yaml
openapi: 3.0.2
info:
  description: |
    Customers focuses on solving authentic identification of humans who are legally able to hold and transfer currency within the US. Primarily this project solves [Know Your Customer](https://en.wikipedia.org/wiki/Know_your_customer) (KYC), [Customer Identification Program](https://en.wikipedia.org/wiki/Customer_Identification_Program) (CIP), [Office of Foreign Asset Control](https://www.treasury.gov/about/organizational-structure/offices/Pages/Office-of-Foreign-Assets-Control.aspx) (OFAC) checks and verification workflows to comply with US federal law and ensure authentic transfers. Customers has an objective to be a service for detailed due diligence on individuals and companies for Financial Institutions and services in a modernized and extensible way.

    Customer phone numbers and addresses are stored and partially used in KYC/OFAC validation. Arbitrary key/value pairs can be stored for a Customer. Documents and Disclaimers, and their acknowledgement are also stored under a Customer as they're accepted. Bank Accounts, which can be validated with micro-deposits currently, are stored under each Customer.

    ![](https://raw.githubusercontent.com/adamdecaf/customers/create-accounts/docs/images/customer.png)
  version: v1
  title: Customers API
  contact:
    url: https://github.com/moov-io/customers
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8087
...
paths:
...
  /customers/{customerID}:
    get:
      tags: [Customers]
      summary: Retrieve customer
      description: Get the Customer object and metadata for the customerID.
      operationId: getCustomer
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional requestID allows application developer to trace requests through the systems logs
          example: rs4f9915
          schema:
            type: string
        - name: X-User-ID
          in: header
          description: Unique userID set by an auth proxy or client to identify and isolate objects.
          example: e3cdf999
          schema:
            type: string
        - name: customerID
          in: path
          description: customerID that identifies this Customer
          required: true
          schema:
            type: string
            example: e210a9d6-d755-4455-9bd2-9577ea7e1081
      responses:
        '200':
          description: A customer objects for the supplied customerID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '404':
          description: No Customer with the specified customerID was found
...
  /customers/{customerID}/status:
    put:
      tags: [Customers]
      summary: Update customer status
      description: Update the status for a customer, which can only be updated by authenticated users with permissions.
      operationId: updateCustomerStatus
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional requestID allows application developer to trace requests through the systems logs
          example: rs4f9915
          schema:
            type: string
        - name: X-User-ID
          in: header
          description: Unique userID set by an auth proxy or client to identify and isolate objects.
          example: e3cdf999
          schema:
            type: string
        - name: customerID
          in: path
          description: customerID of the Customer to update the CustomerStatus
          required: true
          schema:
            type: string
            example: e210a9d6-d755-4455-9bd2-9577ea7e1081
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCustomerStatus'
      responses:
        '200':
          description: A customer object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '400':
          description: Customer status was not updated, see error(s)
          content:
            application/json:
              schema:
                $ref: 'https://raw.githubusercontent.com/moov-io/api/master/openapi-common.yaml#/components/schemas/Error'
  /customers/{customerID}/accounts:
    get:
      tags: [Customers]
      summary: Get Customer Accounts
      description: Retrieve all accounts for the given customer.
      operationId: getCustomerAccounts
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional requestID allows application developer to trace requests through the systems logs
          example: rs4f9915
          schema:
            type: string
        - name: X-User-ID
          in: header
          description: Unique userID set by an auth proxy or client to identify and isolate objects.
          example: e3cdf999
          schema:
            type: string
        - name: customerID
          in: path
          description: customerID of the Customer to get Accounts for
          required: true
          schema:
            type: string
            example: e210a9d6-d755-4455-9bd2-9577ea7e1081
      responses:
        '200':
          description: Active accounts for the customer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Accounts'
        '400':
          description: Failed to get accounts, see error(s)
          content:
            application/json:
              schema:
                $ref: 'https://raw.githubusercontent.com/moov-io/api/master/openapi-common.yaml#/components/schemas/Error'
    post:
      tags: [Customers]
      summary: Create Customer Account
      description: Create an account for the given customer
      operationId: createCustomerAccount
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional requestID allows application developer to trace requests through the systems logs
          example: rs4f9915
          schema:
            type: string
        - name: X-User-ID
          in: header
          description: Unique userID set by an auth proxy or client to identify and isolate objects.
          example: e3cdf999
          schema:
            type: string
        - name: customerID
          in: path
          description: customerID of the Customer to add an Account onto
          required: true
          schema:
            type: string
            example: e210a9d6-d755-4455-9bd2-9577ea7e1081
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccount'
      responses:
        '200':
          description: Account created for the customer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          description: Failed to get accounts, see error(s)
          content:
            application/json:
              schema:
                $ref: 'https://raw.githubusercontent.com/moov-io/api/master/openapi-common.yaml#/components/schemas/Error'
    delete:
      tags: [Customers]
      summary: Delete Customer Account
      description: Remove an account from the given Customer
      operationId: deleteCustomerAccount
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional requestID allows application developer to trace requests through the systems logs
          example: rs4f9915
          schema:
            type: string
        - name: X-User-ID
          in: header
          description: Unique userID set by an auth proxy or client to identify and isolate objects.
          example: e3cdf999
          schema:
            type: string
        - name: customerID
          in: path
          description: customerID of the Customer to remove an Account
          required: true
          schema:
            type: string
            example: e210a9d6-d755-4455-9bd2-9577ea7e1081
      responses:
        '200':
          description: Account removed form customer
        '400':
          description: Failed to get accounts, see error(s)
          content:
            application/json:
              schema:
                $ref: 'https://raw.githubusercontent.com/moov-io/api/master/openapi-common.yaml#/components/schemas/Error'
...
  /customers/{customerID}/ofac:
    get:
      tags: [Customers]
      summary: Get latest OFAC search
      description: Get the latest OFAC search for a customer
      operationId: getLatestOFACSearch
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional requestID allows application developer to trace requests through the systems logs
          example: rs4f9915
          schema:
            type: string
        - name: X-User-ID
          in: header
          description: Unique userID set by an auth proxy or client to identify and isolate objects.
          example: e3cdf999
          schema:
            type: string
        - name: customerID
          in: path
          description: customerID of the Customer to get latest OFAC search
          required: true
          schema:
            type: string
            example: e210a9d6
      responses:
        '200':
          description: OFAC search did not block customer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OFACSearch'
        '400':
          description: An error occurred when refreshing OFAC data, see error(s)
          content:
            application/json:
              schema:
                $ref: 'https://raw.githubusercontent.com/moov-io/api/master/openapi-common.yaml#/components/schemas/Error'
  /customers/{customerID}/refresh/ofac:
    put:
      tags: [Customers]
      summary: Refresh customer OFAC search
      description: Refresh OFAC search for a given Customer
      operationId: refreshOFACSearch
      parameters:
        - name: X-Request-ID
          in: header
          description: Optional requestID allows application developer to trace requests through the systems logs
          example: rs4f9915
          schema:
            type: string
        - name: X-User-ID
          in: header
          description: Unique userID set by an auth proxy or client to identify and isolate objects.
          example: e3cdf999
          schema:
            type: string
        - name: customerID
          in: path
          description: customerID of the Customer to refresh OFAC search
          required: true
          schema:
            type: string
            example: e210a9d6
      responses:
        '200':
          description: OFAC search did not block customer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OFACSearch'
        '400':
          description: An error occurred when refreshing OFAC data, see error(s)
          content:
            application/json:
              schema:
                $ref: 'https://raw.githubusercontent.com/moov-io/api/master/openapi-common.yaml#/components/schemas/Error'
components:
  schemas:
...
    Customer:
      type: object
      properties:
        customerID:
          type: string
          description: The unique identifier for the customer who owns the account
          example: e210a9d6
        firstName:
          type: string
          maximum: 64
          description: Given Name or First Name
          example: Robert
        middleName:
          type: string
          maximum: 64
          description: Middle Name
          example: Flex
        lastName:
          type: string
          description: Surname or Last Name
          example: Smith
        nickName:
          type: string
          description: Name Customer is preferred to be called
          example: Bob
        suffix:
          type: string
          description: Customers name suffix. "Jr", "PH.D."
        type:
          $ref: '#/components/schemas/CustomerType'
        birthDate:
          type: string
          format: date-time
          description: Legal date of birth
          example: '2016-08-29T09:12:33.001Z'
        status:
          $ref: '#/components/schemas/CustomerStatus'
        email:
          type: string
          maximum: 500
          description: Primary email address of customer name@domain.com
        phones:
          type: array
          items:
            $ref: '#/components/schemas/Phone'
        addresses:
          type: array
          items:
            $ref: '#/components/schemas/CustomerAddress'
        metadata:
          type: object
          description: Map of unique keys associated to values to act as foreign key relationships or arbitrary data associated to a Customer.
          additionalProperties:
            type: string
          example:
            paygateID: "23beb5fd"
        createdAt:
          type: string
          format: date-time
          example: '2016-08-29T09:12:33.001Z'
        lastModified:
          type: string
          format: date-time
          description: Last time the object was modified
          example: '2016-08-29T09:12:33.001Z'
    CustomerType:
      type: string
      description: Note if this Customer represents an individual or business
      enum:
        - Individual
        - Business
    CustomerStatus:
      type: string
      description: State of the customer
      enum:
        - Deceased
        - Rejected
        - Unknown
        - ReceiveOnly
        - Verified
...
    UpdateCustomerStatus:
      properties:
        comment:
          type: string
          description: Free form comment about the customer status update
          example: Customer was approved from KYC confirmation
        status:
          $ref: '#/components/schemas/CustomerStatus'
      required:
        - status
...
    Phone:
      properties:
        number:
          type: string
          description: phone number
          example: "+1.818.555.1212"
        valid:
          type: boolean
          description: phone number has been validated to connect with customer
        type:
          type: string
          enum:
            - Home
            - Mobile
            - Work
...
    CustomerAddress:
      type: object
      properties:
        addressID:
          type: string
          description: Unique identifier for this Address
          example: 851233a1
        type:
          type: string
          enum: # TODO(adam): What about adding Previous?
            - Primary
            - Secondary
        address1:
          type: string
          description: First line of the address
        address2:
          type: string
          description: Second line of the address
        city:
          type: string
        state:
          type: string
          minimum: 2
          maximum: 2
          description: two charcer code of US state
        postalCode:
          type: string
        country:
          type: string
          enum:
            - US
        validated:
          type: boolean
          description: Address has been validated for customer
    AccountStatus:
      type: string
      description: Enumeration if this account has been validated through micro-deposits or another available strategy.
      enum:
        - none
        - validated
      example: microdeposits
    AccountType:
      type: string
      description: |
        Defines the account as *checking* or *savings*
      enum:
        - Checking
        - Savings
      example: checking
    CreateAccount:
      properties:
        accountNumber:
          type: string
          description: The account number for the depository account
          example: "0001027028"
        routingNumber:
          type: string
          description: The ABA routing transit number for the depository account.
          example: "051504597"
        type:
          $ref: '#/components/schemas/AccountType'
    Accounts:
      type: array
      items:
        $ref: '#/components/schemas/Account'
    Account:
      properties:
        accountID:
          type: string
          description: A unique identifier for this account
          example: e1b1544a
        maskedAccountNumber:
          type: string
          description: The masked account number for the depository account
          example: "0001027028"
        routingNumber:
          type: string
          description: The ABA routing transit number for the depository account.
          example: "051504597"
        status:
          $ref: '#/components/schemas/AccountStatus'
        type:
          $ref: '#/components/schemas/AccountType'
...
    OFACSearch:
      type: object
      properties:
        entityID:
          type: string
          description: SDN EntityID of the Entity
          example: "2141"
        sdnName:
          type: string
          description: Name of the SDN entity
          example: John Smith
        sdnType:
          type: string
          description: SDN entity type
          example: Individual
        match:
          type: number
          example: 0.91
          description: Percentage of similarity between the Customer name and this OFAC entity
        createdAt:
          type: string
          format: date-time
          example: '2016-08-29T09:12:33.001Z'